Rename data namespace to model
