+++ Epsilon.Canvas.Abstractions/Model/Assignment.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.ObjectModel;
 using System.Text.Json.Serialization;
 
-namespace Epsilon.Canvas.Abstractions.Data;
+namespace Epsilon.Canvas.Abstractions.Model;
 
 public record Assignment(
     [property: JsonPropertyName("id")] int Id,
