+++ Epsilon.Canvas.Abstractions/Model/ModuleItem.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Epsilon.Canvas.Abstractions.Data;
+namespace Epsilon.Canvas.Abstractions.Model;
 
 public record ModuleItem(
     [property: JsonPropertyName("id")] int Id,
