+++ Epsilon.Canvas.Abstractions/Model/Outcome.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Epsilon.Canvas.Abstractions.Data;
+namespace Epsilon.Canvas.Abstractions.Model;
 
 public record Outcome(
     [property: JsonPropertyName("id")] int Id,
