+++ Epsilon.Canvas.Abstractions/Model/Submission.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Epsilon.Canvas.Abstractions.Data;
+namespace Epsilon.Canvas.Abstractions.Model;
 
 public record Submission(
     [property: JsonPropertyName("full_rubric_assessment")] RubricAssessment? RubricAssessment,
