+++ Epsilon.Canvas/Service/AssignmentService.cs
@@ -1,4 +1,4 @@
-﻿using Epsilon.Canvas.Abstractions.Data;
+﻿using Epsilon.Canvas.Abstractions.Model;
 using Epsilon.Canvas.Abstractions.Services;
 using Epsilon.Http.Abstractions;
 using Epsilon.Http.Abstractions.Json;
