+++ Epsilon.Canvas/Service/OutcomeService.cs
@@ -1,4 +1,4 @@
-﻿using Epsilon.Canvas.Abstractions.Data;
+﻿using Epsilon.Canvas.Abstractions.Model;
 using Epsilon.Canvas.Abstractions.Services;
 using Epsilon.Canvas.Response;
 using Epsilon.Http.Abstractions;
